+++ uniScheduleBot/src/main/java/com/example/unischedulebot/exception/SheetsException.java
@@ -1,0 +1,7 @@
+package com.example.unischedulebot.exception;
+
+public class SheetsException extends Exception {
+    public SheetsException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
