+++ uniScheduleBot/src/main/java/com/example/unischedulebot/google/GoogleProps.java
@@ -1,0 +1,15 @@
+package com.example.unischedulebot.google;
+
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Component
+@Getter
+@Setter
+@ConfigurationProperties(prefix = "google")
+public class GoogleProps {
+    private String sheetId;
+}
