+++ uniScheduleBot/src/main/java/com/example/unischedulebot/service/BotService.java
@@ -1,0 +1,10 @@
+package com.example.unischedulebot.service;
+
+import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
+import org.telegram.telegrambots.meta.api.objects.Update;
+
+public interface BotService {
+    SendMessage callbackQueryHelper(Update update);
+
+    SendMessage buttonHelper(Update update);
+}
